+++ NEWS.md
@@ -1,7 +1,7 @@
 # arulesCBA 1.2.6 (2023-xx-xx)
 
 * RCAR: Classifiers with a single rule now work
-* RCAR: Only rules where also the RHS covers the transaction are used.
+* prepareTransactions gained parameter logical2factor.
 
 # arulesCBA 1.2.5 (2022-08-19)